Converted to distance follower and continued conversion to 2019 robot

+++ src/main/java/org/hotteam67/HotPathFollower.java
@@ -80,7 +80,7 @@ public class HotPathFollower
      * @param ticksPerRev
      * @param wheelDiameter
      */
-    public HotPathFollower(int ticksPerRev, double wheelDiameter, String leftPathFile, String rightPathFile)
+    public HotPathFollower(double ticksPerRev, double wheelDiameter, String leftPathFile, String rightPathFile)
     {
         this.ticksPerRev = ticksPerRev;
         this.wheelDiameter = wheelDiameter;
